Made score work properly + bug fixes
I was using reactor not reactorpile so empty reactor points weren't added.
I made utilscript not do a click check if the game is over.
I made resart in the pause menu restart the game instead of going to the menu.
I added config to main menu so we can set difficulty at the start of the game.