Optimizations and a bug fix
Cards and their holograms now have their parent transforms set to their containers.
Holograms are now delt with only when they need to be.
Other card and holograms fixes and improvements.
Cards from the wastepile can no longer be put on a foundation base when it has cards on top.