Optimized saving states
As saving the state is done very often (on every move) and is the most compute intensive function we've added, optimizing this is important.
Combined repetitive actions and removed some unnecessary code. Redid some code so that reversing lists was slightly faster.
The next step is to try using arrays instead of lists and switch from foreach to for loops. A smarter saving feature could also be implemented through editing the previous save state instead of remaking them.