Added Failsafe for Loading Saved Game
Users can now be greeted with an popup explaining that their save failed to load. The save is deleted if this occurs.
Made changes to game play loading to support this.
The settings popup window was made into a prefab. Its buttons now have the normal sound and vibration effects.