Refactored Sorting Layers and Conveyor Belt
Divided the gameplay layer into a background and foreground.
Added a card layer in-between the gameplay layers.
As a result, updated positioning in the gameplay scene and some prefabs.
Updated position compensations made in-script.

Remade the conveyor belt so that its positioning wasn't flipped.
Replaced the large belt sprite asset with a smaller one.