Added Hidden Cheat Mode
Activating cheat mode loads a special difficulty that includes a lowered foundation starting size, increased reactor limits, and a reduced move count. All game wins are counted as if all cards were matched for the summary animation.
Everything else is similar and works like the other difficulties.
This mode is accessed by pressing on the space ship in the main menu and can only be exited by returning to the main menu.
This is primarily to be used for playtesting purposes.