Waste Pile Animation Improvements and Minor Fix
Removed the slight visual stutter on card removals. This was caused by the slight offsetting of belt card containers causing misalignments for the belt image pieces. Added an additional belt piece that is always to the left and adjusted all belt positionings to account for the offset.
When the waste pile is empty and cards are added to it, they are now moved offscreen before scrolling into view.
Also fixed the game not automatically ending when it should in certain cases.