Added WebGL dual build support for desktop and mobile
This adds a menu item it the Unity Editor that triggers a dual build.
This build creates two data packs for each platform to support their data compression capabilities.
Also includes a custom WebGL template with minor modifications and the ability to load the correct data pack depending on the platform.