Updated card movement and auto placement time settings
Card movement has been tweaked to end closer to the target position, so all speeds had to be reduced to compensate.
The auto placement time input allows for formats of either "000" "0.0" or ".00" to be entered but it cannot display the ".00" format automatically. To remove that option the input has been limited further and rounded.